sync/requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
Markdown==3.8.2
PyYAML==6.0.2
requests==2.32.4
urllib3==2.5.0
